pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "flask-blog-generator"
version = "0.1.0"
description = "static blog generator with flask"
authors = [
    {name = "Jessie Anderson",email = "housewifehacker@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "python-frontmatter (>=1.1.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
